app/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
datasets
accelerate
sentencepiece
sqlalchemy
alembic
pytest
pytest-asyncio
python-dotenv
psycopg2-binary
python-multipart
PyPDF2
python-docx
google-generativeai